refactor: Constant 수정
